+++ idea-plugin/src/test/data/libFixtures/kotlin/metro/hints/hints.kt
@@ -81,3 +81,7 @@ fun libtest_LibMultibindingScope(
 ) {}
 
 internal fun dev_zacsweers_metro_AppScope(contributed: LibHiddenImpl) {}
+
+fun libtest_LibOriginScope(contributed: libtest.LibOriginOriginalBindings) {}
+
+fun libtest_LibOriginScope(contributed: libtest.LibOriginReplacementProviders.ToScope) {}